Move the getUserSettings function from auth to user-managment service
